Ensure packages within a Package Set are unique
To illustrate this functioning, forced `img.yml` to overlay on the
Package Set supplied by the "common" layer... and include both a new
package ("coredns") and an existing package ("vim").

This is achieved by annotating every package included in a Package Set
with a matcher that first searches the array for the same named package.
If that matcher does _not_ find a match (i.e. it is "missing"), that's
okay, just perform the default behavior, then; and the default behavior
of overlaying array items is to append the item to the end of the array.