Render each layer as a ytt invocation, overlaying from the last (#1)
* Render each layer as a ytt invocation, overlaying from the last

In the Makefile:
- the basis of an image is to build the "base" layer
- "common" _overlays_ on top of that (and yes, overlays can be
  templatized)
- "img" _overlays_ on top of that
- cross-layer switches (like DISTRO) are Makefile variables

In each layer:
- collect all ytt files (except values) into a single directory; ytt
  will sort out the correct order of files.
- for "base" and "common" define per-distro values.

* Extract "Package Set" YAML as a shared function

- these are being contributed by both the "common" and "img" layers.
- Package Sets appear to require an "action"
  - it was missed in the example in the "common" layer
  - it is most often "install" which makes it somewhat boilerplate

* Pass structure of "source" data value in "base"

Illustrates how a data value itself can be a whole substructure.

Practically, this means that portions of YAML can be strictly enforced
through Data Value Schema.

* Ensure packages within a Package Set are unique

To illustrate this functioning, forced `img.yml` to overlay on the
Package Set supplied by the "common" layer... and include both a new
package ("coredns") and an existing package ("vim").

This is achieved by annotating every package included in a Package Set
with a matcher that first searches the array for the same named package.
If that matcher does _not_ find a match (i.e. it is "missing"), that's
okay, just perform the default behavior, then; and the default behavior
of overlaying array items is to append the item to the end of the array.

* (minor) Rename "lib" to "config"; to be more traditional

It's been a minute and I forgot that "config" was the more "standard" or
traditional name for the directory containing all `ytt` files _except_
the data values.